+++ src/Resources/ResourceManager/Implementation/DeploymentStacks/RemoveAzManagmentGroupDeploymentStack.cs
@@ -100,14 +100,20 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
 
                 string confirmationMessage = $"Are you sure you want to remove ManagementGroup scoped DeploymentStack '{Name}' " +
                     $"in ManagementGroup '{ManagementGroupId}' with the following actions?" +
-                    (!shouldDeleteResources || !shouldDeleteResourceGroups ? "\nDetaching: " : "") +
+                    // Deleting
+                    (shouldDeleteResources || shouldDeleteResourceGroups || shouldDeleteManagementGroups ? "\nDeleting: " : "") +
+                    (shouldDeleteResources ? "resources" : "") +
+                    (shouldDeleteResources && shouldDeleteResourceGroups ? ", " : "") +
+                    (shouldDeleteResourceGroups ? "resourceGroups" : "") +
+                    (shouldDeleteResourceGroups && shouldDeleteManagementGroups ? ", " : "") +
+                    (shouldDeleteManagementGroups ? "managementGroups" : "") +
+                    // Detaching
+                    (!shouldDeleteResources || !shouldDeleteResourceGroups || !shouldDeleteManagementGroups ? "\nDetaching: " : "") +
                     (!shouldDeleteResources ? "resources" : "") +
                     (!shouldDeleteResources && !shouldDeleteResourceGroups ? ", " : "") +
                     (!shouldDeleteResourceGroups ? "resourceGroups" : "") +
-                    (shouldDeleteResources || shouldDeleteResourceGroups ? "\nDeleting: " : "") +
-                    (shouldDeleteResources ? "resources" : "") +
-                    (shouldDeleteResources && shouldDeleteResourceGroups ? ", " : "") +
-                    (shouldDeleteResourceGroups ? "resourceGroups" : "");
+                    (!shouldDeleteResourceGroups && !shouldDeleteManagementGroups ? ", " : "") +
+                    (!shouldDeleteManagementGroups ? "managementGroups" : "");
 
                 ConfirmAction(
                     Force.IsPresent,